Move field monkeypatch to sphinx_pyodide
