+++ docs/sphinx_pyodide/sphinx_pyodide/__init__.py
@@ -42,8 +42,26 @@ def patch_templates():
     JsRenderer.rst = patched_rst_method
 
 
+def patch_field():
+    """
+    Add extra CSS classes to some documentation fields. We use this in
+    pyodide.css to fix the rendering of autodoc return value annotations.
+    """
+    from sphinx.util.docfields import Field
+
+    orig_make_field = Field.make_field
+
+    def make_field(self, *args, **kwargs):
+        node = orig_make_field(self, *args, **kwargs)
+        node["classes"].append(self.name)
+        return node
+
+    Field.make_field = make_field
+
+
 def setup(app):
     patch_templates()
+    patch_field()
     monkeypatch_module_documenter()
     app.add_lexer("pyodide", PyodideLexer)
     app.add_lexer("html-pyodide", HtmlPyodideLexer)
